Refactor: split into workspaces
This is mostly a straightforward refactoring, the reduction in line
count is because the new `dns-types` library has very few
dependencies.
