+++ bin-resolved/Cargo.toml
@@ -1,0 +1,23 @@
+[package]
+name = "resolved"
+version = "0.1.0"
+edition = "2021"
+
+# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
+
+[dependencies]
+async-recursion = "1"
+bytes = "1"
+clap = { version = "3", features = ["derive"] }
+dns-types = { path = "../lib-dns-types" }
+priority-queue = "1"
+rand = "0.8.5"
+tokio = { version = "1", features = ["rt-multi-thread", "net", "macros", "io-util", "sync", "fs", "time"] }
+
+[dev-dependencies]
+criterion = "0.3.5"
+dns-types = { path = "../lib-dns-types", features = ["test-util"] }
+
+[[bench]]
+name = "cache"
+harness = false
