+++ bin-resolved/src/lib.rs
@@ -1,0 +1,3 @@
+pub mod fs_util;
+pub mod net_util;
+pub mod resolver;
