+++ bin-resolved/src/net_util.rs
