+++ fuzz/fuzz_targets/wire_serialise_round_trip.rs
@@ -1,7 +1,7 @@
 #![no_main]
 use libfuzzer_sys::fuzz_target;
 
-use resolved::protocol::types::Message;
+use dns_types::protocol::types::Message;
 
 fuzz_target!(|message: Message| {
     let serialised = message.clone().to_octets().unwrap();
