+++ fuzz/fuzz_targets/zone_deserialise_round_trip.rs
@@ -1,7 +1,7 @@
 #![no_main]
 use libfuzzer_sys::fuzz_target;
 
-use resolved::zones::types::Zone;
+use dns_types::zones::types::Zone;
 
 fuzz_target!(|data: &str| {
     if let Ok(deserialised) = Zone::deserialise(data) {
