+++ lib-dns-types/src/hosts/serialise.rs
