+++ lib-dns-types/src/lib.rs
@@ -1,5 +1,3 @@
 pub mod hosts;
-pub mod net_util;
 pub mod protocol;
-pub mod resolver;
 pub mod zones;
