+++ lib-dns-types/src/protocol/types.rs
@@ -19,7 +19,7 @@ use std::net::{Ipv4Addr, Ipv6Addr};
 ///
 /// See section 4.1 of RFC 1035.
 #[derive(Debug, Clone, Eq, PartialEq, Ord, PartialOrd, Hash)]
-#[cfg_attr(any(feature = "arbitrary", test), derive(arbitrary::Arbitrary))]
+#[cfg_attr(any(feature = "test-util", test), derive(arbitrary::Arbitrary))]
 pub struct Message {
     pub header: Header,
     pub questions: Vec<Question>,
@@ -109,7 +109,7 @@ impl Message {
 ///
 /// See section 4.1.1 of RFC 1035.
 #[derive(Debug, Copy, Clone, Eq, PartialEq, Ord, PartialOrd, Hash)]
-#[cfg_attr(any(feature = "arbitrary", test), derive(arbitrary::Arbitrary))]
+#[cfg_attr(any(feature = "test-util", test), derive(arbitrary::Arbitrary))]
 pub struct Header {
     /// A 16 bit identifier assigned by the program that generates any
     /// kind of query.  This identifier is copied the corresponding
@@ -194,7 +194,7 @@ pub struct Header {
 /// in the normal `Header` type would require ensuring those values
 /// are correct.
 #[derive(Debug, Copy, Clone, Eq, PartialEq, Ord, PartialOrd, Hash)]
-#[cfg_attr(any(feature = "arbitrary", test), derive(arbitrary::Arbitrary))]
+#[cfg_attr(any(feature = "test-util", test), derive(arbitrary::Arbitrary))]
 pub struct WireHeader {
     /// The header that will be persisted to / is taken from the
     /// `Message`.
@@ -237,7 +237,7 @@ pub struct WireHeader {
 ///
 /// See section 4.1.2 of RFC 1035.
 #[derive(Debug, Clone, Eq, PartialEq, Ord, PartialOrd, Hash)]
-#[cfg_attr(any(feature = "arbitrary", test), derive(arbitrary::Arbitrary))]
+#[cfg_attr(any(feature = "test-util", test), derive(arbitrary::Arbitrary))]
 pub struct Question {
     /// a domain name represented as a sequence of labels, where each
     /// label consists of a length octet followed by that number of
@@ -292,7 +292,7 @@ impl Question {
 ///
 /// See section 4.1.3 of RFC 1035.
 #[derive(Debug, Clone, Eq, PartialEq, Ord, PartialOrd, Hash)]
-#[cfg_attr(any(feature = "arbitrary", test), derive(arbitrary::Arbitrary))]
+#[cfg_attr(any(feature = "test-util", test), derive(arbitrary::Arbitrary))]
 pub struct ResourceRecord {
     /// a domain name to which this resource record pertains.
     pub name: DomainName,
@@ -602,7 +602,7 @@ impl RecordTypeWithData {
     }
 }
 
-#[cfg(any(feature = "arbitrary", test))]
+#[cfg(any(feature = "test-util", test))]
 impl<'a> arbitrary::Arbitrary<'a> for RecordTypeWithData {
     // this is pretty verbose but it feels like a better way to
     // guarantee the max size of the `Vec<u8>`s than adding a wrapper
@@ -711,7 +711,7 @@ impl From<Opcode> for u8 {
     }
 }
 
-#[cfg(any(feature = "arbitrary", test))]
+#[cfg(any(feature = "test-util", test))]
 impl<'a> arbitrary::Arbitrary<'a> for Opcode {
     fn arbitrary(u: &mut arbitrary::Unstructured<'a>) -> arbitrary::Result<Self> {
         Ok(Self::from(u.arbitrary::<u8>()?))
@@ -769,7 +769,7 @@ impl From<Rcode> for u8 {
     }
 }
 
-#[cfg(any(feature = "arbitrary", test))]
+#[cfg(any(feature = "test-util", test))]
 impl<'a> arbitrary::Arbitrary<'a> for Rcode {
     fn arbitrary(u: &mut arbitrary::Unstructured<'a>) -> arbitrary::Result<Self> {
         Ok(Self::from(u.arbitrary::<u8>()?))
@@ -912,7 +912,7 @@ impl fmt::Debug for DomainName {
     }
 }
 
-#[cfg(any(feature = "arbitrary", test))]
+#[cfg(any(feature = "test-util", test))]
 impl<'a> arbitrary::Arbitrary<'a> for DomainName {
     fn arbitrary(u: &mut arbitrary::Unstructured<'a>) -> arbitrary::Result<Self> {
         let num_labels = u.int_in_range::<usize>(0..=10)?;
@@ -988,7 +988,8 @@ impl From<QueryType> for u16 {
         }
     }
 }
-#[cfg(any(feature = "arbitrary", test))]
+
+#[cfg(any(feature = "test-util", test))]
 impl<'a> arbitrary::Arbitrary<'a> for QueryType {
     fn arbitrary(u: &mut arbitrary::Unstructured<'a>) -> arbitrary::Result<Self> {
         Ok(Self::from(u.arbitrary::<u16>()?))
@@ -1029,7 +1030,7 @@ impl From<QueryClass> for u16 {
     }
 }
 
-#[cfg(any(feature = "arbitrary", test))]
+#[cfg(any(feature = "test-util", test))]
 impl<'a> arbitrary::Arbitrary<'a> for QueryClass {
     fn arbitrary(u: &mut arbitrary::Unstructured<'a>) -> arbitrary::Result<Self> {
         Ok(Self::from(u.arbitrary::<u16>()?))
@@ -1153,7 +1154,7 @@ impl From<RecordType> for u16 {
     }
 }
 
-#[cfg(any(feature = "arbitrary", test))]
+#[cfg(any(feature = "test-util", test))]
 impl<'a> arbitrary::Arbitrary<'a> for RecordType {
     fn arbitrary(u: &mut arbitrary::Unstructured<'a>) -> arbitrary::Result<Self> {
         Ok(Self::from(u.arbitrary::<u16>()?))
@@ -1203,7 +1204,7 @@ impl From<RecordClass> for u16 {
     }
 }
 
-#[cfg(any(feature = "arbitrary", test))]
+#[cfg(any(feature = "test-util", test))]
 impl<'a> arbitrary::Arbitrary<'a> for RecordClass {
     fn arbitrary(u: &mut arbitrary::Unstructured<'a>) -> arbitrary::Result<Self> {
         Ok(Self::from(u.arbitrary::<u16>()?))
@@ -1378,7 +1379,7 @@ mod tests {
     }
 }
 
-#[cfg(test)]
+#[cfg(any(feature = "test-util", test))]
 pub mod test_util {
     use super::*;
